removed extra files and wrote gene module into tfac
